0.6 feature dev (#25)
see changelog for 0.6

+++ src/main/java/renderer/buffers/LineRenderQueue.java
@@ -88,7 +88,10 @@ public abstract class LineRenderQueue
 
 	public static void delete()
 	{
-		lines.delete();
+		if (lines != null) {
+			lines.delete();
+			lines = null;
+		}
 	}
 
 	public static void printContents()
